requirements.txt
python-telegram-bot
pydantic
aiohttp